requirements.txt
groq~=0.19.0
flask~=3.0.0
functions~=0.7.0
openai>=1.3.0
packaging~=23.2
requests~=2.31.0
waitress
python-dotenv==1.0.1